Quick fix to get BE running with Roddy.
Problem:

Previously CONFIG_FILE and PARAMETER_FILE were only used in PBS, so LSF did not work. The ideal solution is to leave the definition and handling of parameter files to the client code or to come up with a clear interface for using parameter files. For the former case it is currently not trivial to remove the parameters from the job (one would have to either create a temporary job object with reduced parameters, or allow explicit setting of parameters during job running. The code is currently not well prepared for this). The latter solution has some unresolved questions: How to set the name of the parameter file variable? Where to save the files (tmp, some notion of execution directory, ...)?