Changes for review of #137 by vinjana
Update gradle wrapper to 5.1.1
Updated the dependency for RoddyToolLib to 2.1.0

In GenericJobInfo, change field tool of type File to command of type
String. This better reflects that e.g. LSF can also execute commands. A
tool can always be seen as a command whereas a command is not
necessarily a tool.

In JobManagerOptions, rename maxAgeOfJobsForJobQueries to
maxTrackingTimeForFinishedJobs, as the new name is much more expressive.

In BatchEuphoriaJobManager and its subclasses, rename the aforementioned
field as well.

In PBSCommandParser and LSFCommandParser, do a bit of code cleanup and
adapt code.

In LSFJobManager:
- Add the stripAwayStatusInfo method for time string. This gets rid of
  the trailing status code in some timing info reported by LSF.
- Change the return value of runBjobs to return a nested map of type
  Map<String,String>. This makes some followup code much cleaner.
- Changed convertJobDetailsMapToGenericJobInfoobject a lot:
  - Fixed a typo.
  - Got rid of all the "as String" casts.
  - Reordered the entries and grouped them logically.
  - Extracted code duplicates to custom methods.

In LSFJobManagerSpec:
- Renamed "queryJobInfo, bjobs JSON output with lists" to
          "Test convertJobDetailsMapToGenericJobInfoObject"
- Added the "test queryExtendedJobStateById with overdue date" method
  which tests for a NullPointerException.